Adding Map Reload Feature
Problem Description:
Sometimes map does not load or loads incomplete, due to lack of internet connectivity or some other issue.

Which is why, I'm adding the feature to reload the map.

Currently, the map reloads when clicking on the yatri-logo at the bottom. Later a refresh button will be added.